#136: チャットユーザー選択画面、dataSource作成
